==> notebooks/01_entpacken_und_rohdatenanalyse.ipynb <==
### Entpacken und Analyse der Rohdaten

Dieses Notebook dient der Vorbereitung und ersten Analyse der verfügbaren Rohdaten.  
Ziel ist es, die gelieferten Datenarchive (ZIP/TAR) zu entpacken, die enthaltenen HTML- und CSV-Dateien im Data Lake abzulegen und die Datenqualität zu prüfen.

Die Ergebnisse dieses Notebooks bilden die Grundlage für die spätere inhaltliche Verarbeitung in 02_verarbeitung_main.ipynb.

#### 1. Import benötigte Pakete

# Standard
import os # Dateipfaden
import pandas as pd # Tabellenverarbeitung (DataFrames)
from datetime import datetime # Datumsangaben

# Datenprofiling
from ydata_profiling import ProfileReport  

# Dateiverwaltung & Entpacken
from glob import glob # Mehrere Dateien suchen
import tarfile # .tar.gz Dateien entpacken

# Pfade
# Projektverzeichnis 
PROJECT_ROOT = r"D:/DBU/ADSC11 ADS-01/Studienarbeit/newspaper-scraping"

# Input-Pfade
INPUT_PATH = os.path.join(PROJECT_ROOT, "input", "raw") # Rohdaten
DATA_LAKE_PATH = os.path.join(INPUT_PATH, "data-lake") # HTML- und CSV-Dateien
ZIP_PATH = os.path.join(INPUT_PATH, "downloaded_zips") # ZIP-Dateien

# Output-Pfad
OUTPUT_PATH = os.path.join(PROJECT_ROOT, "output") 

#### 2. Archiv entpacken und Dateien speichern

# Funktion: Mit tarfile Rohdaten (Format .tar.gz) einlesen, extrahieren & entpacken
def extract_tar_file(ZIP_PATH, DATA_LAKE_PATH): 
    entpackt = 0 # zählen
    with tarfile.open(ZIP_PATH, "r:gz") as tar:
        for member in tar.getmembers():
            
            # HTML- und CSV-Dateien extrahieren
            if member.name.endswith(".html") or member.name.endswith(".csv"):
                target_path = os.path.join(DATA_LAKE_PATH, os.path.basename(member.name))
                
                # Entpacken in HTML-Datei
                with open(target_path, "wb") as f:
                    f.write(tar.extractfile(member).read())
                entpackt += 1
    return entpackt  

# Entpackung: Mit glob nach .tar.gz-Dateien im ZIP-Ordner suchen und extrahieren
gesamt_entpackt = 0
for zip_path in glob(os.path.join(ZIP_PATH, "*.tar.gz")):
    gesamt_entpackt += extract_tar_file(zip_path, DATA_LAKE_PATH)

# Ergebnis anzeigen
print(f"\n{gesamt_entpackt} neue Dateien entpackt.")

# Verfügbare HTML- und CSV-Dateien im DATA_LAKE_PATH zählen und zurückgeben
html_files = glob(os.path.join(DATA_LAKE_PATH, "*.html"))
csv_files = [f for f in glob(os.path.join(DATA_LAKE_PATH, "*.csv"))]

print("Anzahl CSV-Dateien:", len(csv_files))
print("Anzahl HTML-Dateien:", len(html_files))

#### 3. Datenexploration

##### 3.1 CSV-Dateien (Metadaten)

# CSV-Dateien mit glob finden
csv_files = glob(os.path.join(DATA_LAKE_PATH, "*.csv"))

# Einzelne CSV-Dateien laden und in Liste speichern
dfs = []
for f in csv_files:
    try:
        df = pd.read_csv(f)
        # leeere Dateien überspringen
        if not df.empty:
            df["filename"] = os.path.basename(f)
            dfs.append(df)
    except Exception as e:
        print(f"[Fehler] Datei konnte nicht gelesen werden: {f} ({e})")

# In DataFrame zusammenführen
df_metadaten = pd.concat(dfs, ignore_index=True)

# Überblick df_metadaten
df_metadaten.head()

# Anzahl Zeilen und Spalten
df_metadaten.shape

# Übersicht Daten
# Data profiling 
profile = ProfileReport(df_metadaten)

# ProfileReport als html
profile.to_file(os.path.join(OUTPUT_PATH, "rohdaten_profile_report.html"))

# Zeitraum 
df_metadaten["date"] = pd.to_datetime(df_metadaten["date"])  

print("Zeitraum:")
print("Von:", df_metadaten["date"].min().date())
print("Bis:", df_metadaten["date"].max().date())

# Überblick Medienquelle
df_metadaten["name"].unique()

# Anzahl Medienquelle
print("Anzahl Medienquellen:", df_metadaten["name"].nunique())

# Anzahl Medienquelle nach Jahr
# Jahr extrahieren
df_metadaten["year"] = df_metadaten["date"].dt.year

# Gruppieren nach nach Medium und Jahr
df_medien_nach_jahr = df_metadaten.groupby("year")["name"].nunique().reset_index()
df_medien_nach_jahr

# Anzahl Daten pro Medienquelle
df_metadaten["name"].value_counts()

# Übersicht Medienquellen und zugehörige URL
df_medienliste = df_metadaten[["name", "original_url"]].drop_duplicates()
df_medienliste

# Medienliste als CSV exportieren
df_medienliste.to_csv(os.path.join(OUTPUT_PATH, "medien_quellen_uebersicht.csv"))

# Fehlende Werte 
df_metadaten.isna().sum()

# Fehlende Werte beim Encoding 
df_metadaten[df_metadaten["encoding"].isna()][["name", "date", "file_name", "original_url"]]

# Duplikate
df_metadaten.duplicated().sum()

# HTTP-Status
df_metadaten["status"].value_counts()

##### 3.2 HTML-Dateien

# HTML-Dateien mit glob finden
html_files = glob(os.path.join(DATA_LAKE_PATH, "*.html"))

# Dateinamen extrahieren und in Liste speichern
html_data = [] 
for f in html_files:
    name = os.path.basename(f)
    html_data.append({"filename": name})

# Als DataFrame speichern
df_html = pd.DataFrame(html_data)

# Überblick df_html
df_html.head()

# HTML-Dateinamen extrahieren und Jahr ableiten
html_data = []
for f in html_files:
    name = os.path.basename(f)
    year = int(name[:4])  # Jahr direkt aus den ersten 4 Zeichen
    html_data.append({"filename": name, "year": year})

df_html = pd.DataFrame(html_data)

# HTML-Dateien pro Jahr zählen
df_html_jahr = df_html["year"].value_counts().sort_index().reset_index()
df_html_jahr

# HTML-Dateien nach Medium
df_html["medium"] = df_html["filename"].str[11:-5]  # alles zwischen Datum und ".html"
df_html["medium"].value_counts().head(10)

# Ergebnis als CSV exportieren
df_html_medium_counts = df_html["medium"].value_counts().reset_index()
df_html_medium_counts.columns = ["medium", "anzahl_html_dateien"]
df_html_medium_counts.to_csv(os.path.join(OUTPUT_PATH, "html_dateien_pro_medium.csv"))
